balloon_project_modular/requirements.txt
easyocr==1.7.2
numpy==2.2.5
opencv_python==4.10.0.84
Pillow==11.2.1
ultralytics==8.3.124
